Improve legibility on signal handler
